Update data.js
Here, the image property is defined as an object, which does not align with the defined schema for the WANDERLUST Project. Because of that, we may encounter a validation error. So, I've changed the image type from Object to String. I think it will help other students who have not dealt with it yet.